+++ backend/db/models/spot.js
@@ -66,11 +66,11 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     lat: {
-      type: DataTypes.REAL,
+      type: DataTypes.DECIMAL,
       allowNull: false
     },
     lng: {
-      type: DataTypes.REAL,
+      type: DataTypes.DECIMAL,
       allowNull: false
     },
     name: {
@@ -82,7 +82,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     price: {
-      type: DataTypes.REAL,
+      type: DataTypes.DECIMAL,
       allowNull: false
     }
   }, {
